+++ sphinx/application/network/concepts/queries/concept-network-query-people/src/main/java/chat/sphinx/concept_network_query_people/model/BadgeStateDto.kt
@@ -5,5 +5,5 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class BadgeStateDto(
     val badge_id: Int,
-    val chat_id: Long
+    val chat_id: Int
 )
